+++ History.md
@@ -10,6 +10,7 @@
 *   [#40](https://github.com/pmd/pmd-regression-tester/issues/40): NoMethodError on beta3
 *   [#42](https://github.com/pmd/pmd-regression-tester/issues/42): Installing the snapshot version of pmdtester locally from github
 *   [#46](https://github.com/pmd/pmd-regression-tester/issues/46): Support multithreaded execution of PMD
+*   [#47](https://github.com/pmd/pmd-regression-tester/issues/47): Reuse the already built PMD binary for PR checks
 *   [#49](https://github.com/pmd/pmd-regression-tester/issues/49): Support comparing two error stacktraces
 *   [#50](https://github.com/pmd/pmd-regression-tester/issues/50): Differences due to different locale settings
 *   [#57](https://github.com/pmd/pmd-regression-tester/issues/57): Support configuration errors in the report